+++ quiz-map/index.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <link rel="stylesheet" href="../common/style.css">
+    <link rel="stylesheet" href="./quiz-map.css">
+
+    <title>The Quiz</title>
+</head>
+<body>
+    <h1>Pick a topic</h1>
+    <ul id = "quiz-list"></ul>
+    
+</body>
+<script type="module" src="./quiz-map.js"></script>
+</html>
