+++ quiz-map/quiz-map.js
@@ -1,7 +1,23 @@
-//check local storage to see what questions have been completed
+// Import metadata
+import quizzes from '../data/quizData.js';
+import { getUserStatus } from '../common/utils.js';
 
-//if there are no questions left to answer
+
+//check local storage to see what questions have been completed (localStorage.getItem. Located in utils.js. Import getUserStatus above. 
+//call user status to check local storage.
+getUserStatus();
+
+
+//if there are no questions left to answer (loop through quizzes array and check if completed)
+
+
+
+
     //redirect to the results page
+
+
+
+
 
 //if there are any questions left to answer...
 //render a list of REMAINING questions based on the metadata
@@ -16,3 +32,6 @@
             //add the href to the a tag with query params for the question
             //append anchor tag to the li
         //append the question li to the ul
+
+        // Loop through items and create links for each one
+// Use href like quest/id=dragon to "pass" the selected id to the quest page
